add: change the start page
We want to test how users create new forms. Currently the link to the
forms jumps straight in to changing the start page and adding pages.

A new page with an action button has been added, which leads to giving
the form a name, then the page list. The prototype still only supports
creating one form.