add: move the add new page button and add preview
The 'create new page' button on the page-list may not be visible if
there are many pages already added. Now the in page preview has been
removed, there is no way to preview the form from the page-list.

The button has been moved to the top and a link added to open the new
tab preview at the start page.