April ur test improvements (#29)
* Add entry point from management app

* Change fields to fit our first private beta form

* Remove settings page

* fix: remove choose a page type

The choose a page type page has options which are not implemented yet -
use a template, copy another page. These might confuse users when
testing.

Hackily added a redirect to current handler to remove the choose a page
type page

* add: change the start page

We want to test how users create new forms. Currently the link to the
forms jumps straight in to changing the start page and adding pages.

A new page with an action button has been added, which leads to giving
the form a name, then the page list. The prototype still only supports
creating one form.

* fix: remove units from single text line question

Having a units conditional when choosing single text line might confuse
our users in the first review sessions.

Deleted.

* add: change location of preview links/buttons

The placement of 'update preview' and 'open in new tab' might confuse
our users.

They have been made more promiment, moved to the top and given equal
weight. (Horrible CSS to do so as one is a button the other a link...
its a prototype though, right?)

* fix: remove edit start page from page-list

* fix: make sure all pages have a title on page-list

Some pages didn't appear on the page-list, even though they had been
created.

Pages previously used short-title in the page-list. Now they use
short-title or long-title or pageIndex if those haven't been filled in.

* fix: remove edit start page

Might be confusing to our first users.

* fix: remove form preview from page-list

Remove the form preview from the page list to simplify the flow.

* fix: check-your-answers shows all questions

The check your answers page doesn't show questions unless they have a
short question name set.

The page now shows all answers, using long-title and page number if no
short name has been set.

* Tweak preview styling

* Tweak single line answer content

* fix the double deletePage

* fix: stop in page preview changing page on click

The in page preview changes the page being edited if you click
'Continue'.

The attribute 'sandbox' is added to the iframe which contains the
preview to disable all the links.

* fix: correct the publisher link to create new form

* fix: fix update preview when no response type set

When clicking update preview with a title and short name, the preview
does not update.

This is because the question is not rendered when there is no type set.
A default type of text is now set using the radio buttons.

* add: move the add new page button and add preview

The 'create new page' button on the page-list may not be visible if
there are many pages already added. Now the in page preview has been
removed, there is no way to preview the form from the page-list.

The button has been moved to the top and a link added to open the new
tab preview at the start page.

* fix: standard size for text inputs in preview

After removing units, the text input size was reduced.

This removes the units dispatch from the template so it defaults to text
type.

* fix: spelling - from -> form

Co-authored-by: David Biddle <david.biddle@digital.cabinet-office.gov.uk>
Co-authored-by: Tom Iles <tom.iles@engineering.digital.dwp.gov.uk>
Co-authored-by: agz deberny <a@deberny.com>