fix: in-page-preview stopped showing styling
The preview shows unstyled content.

This is because the sandbox attribute was added to the iframe containing
the preview. sandbox prevents the iframe loading content from the same
origin and runninng scripts.

It's been changed to:

sandbox="allow-same-origin allow-scripts"

This doesn't provide much security - in our case we are not worried
about that, so it shouldn't matter unless we stop trusting the content
of the iframe at some point.