Remove short-name input field
Short name has been removed from the edit-question page and is no longer
used in any of the other pages.

The biggest visited change is in the CYA page, which used short name as
a heading for questions to be changed.

Short name was also used on pages with intro text (a feature that has
also been removed).

The best way to add short-name back to the prototype would be to revert
this commit.