Add 'number' answer type
During demonstrations we often have questions which could best be
answered with a number. The prototype originally had a numeric answer
type but it was removed to suport user testing sessions which did not
require it.

A feature flag has been added to enable a number answer type which
behaves like the short-text type.

To test it, start the prototype with the following command:

`ENABLE_NUMBER_ANSWER_TYPE=true npm run start`