Update edit form submitted page design
We have decided to remove editing the title for confirmation page message and keep it general.

Removed reference number as we won't be generating it.

Edit submitted form page has iterated design, where we include form name in the H1 hint text to be consistent.

We've included additional guidance on what this page is for and how form creators could include a useful content for 'What happens next' section.

We have removed default text and hint text in the textarea for 'What happens next'. We gave an example copy in the inset text above, to encourage form creators in providing their own content for this section.