(fix) Update Declaration status tag and ask if section is complete
By making declaration optional we have no way of marking a task as complete or in progress.

So we added a question on the page if the task is complete or if a user wants come back later and finish it.

This helps us to update the status tag from Not started - to - in progress - to - complete

Not entirely sure why make form live wasn't getting active so I've reduced the counter to less or equal 6