Remove store data workaround
This workaround was needed to handle situations where we a form submits
data named something like `pages[<n>][<key>]`. However, in a subsequent
commit (96fbeff), the routes that update the pages list were refactored,
to make it easier for our designers to maintain [[1]]. I believe that
this means that the workaround code in utils is no longer reachable, and
can be removed. I have left a warning in case I'm wrong...

[1]: https://gds.slack.com/archives/C030K5706TX/p1678457780289839?thread_ts=1678444323.339829&cid=C030K5706TX