Fix Markdown being escaped
Post-migration to v13 of the kit, I found that Markdown was being
escaped. It took me a while to figure out what was going on, but I
traced the issue down to a line in the Nunjucks code base, [[1]], that
checks whether a string needs to be escaped or not with `instanceof`.
The MDN docs for `instanceof` note that there can be issues with
checking objects from different 'realms' [[2]].

This is what was happening here; because `govuk-prototype-kit` is
shrinkwrapped, when we installed `nunjucks-markdown` it was unable to
use the copy of `nunjucks` that comes with the kit to satisfy its peer
dependency, and had to install its own copy. With two separate
`nunjucks` modules, the JavaScript virtual machine will have two copies
of the `SafeString` prototype at two different memory locations, with
the `Environment` from the kit expecting one memory location and the
`nunjucks-markdown` using the other from a different memory location.

The workaround for this is fairly straightfoward, luckily; we can just
tell npm to symlink the copy of `nunjucks` that comes with the kit to
our own `node_modules` folder, with

        npm install nunjucks \
                file:node_modules/govuk-prototype-kit/node_modules/nunjucks

This is a little nasty, but it works :smiling_imp:

[1]: https://github.com/mozilla/nunjucks/blob/fd500902d7c88672470c87170796de52fc0f791a/nunjucks/src/runtime.js#L212
[2]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/instanceof#instanceof_and_multiple_realms