Use plugin for nunjucks-markdown
Telling npm to link the nunjucks package to itself was causing issues
with `npm ci`, possibly due to a bug in npm, but also probably due to
the fact that it is not very tree-like.

This commit changes the prototype to use a forked version of
nunjucks-markdown, [lfdebrux/govuk-prototype-kit-nunjucks-markdown-plugin],
which has been tweaked to work the prototype kit so doesn't need the
same workaround.

[lfdebrux/govuk-prototype-kit-nunjucks-markdown-plugin]: https://github.com/lfdebrux/govuk-prototype-kit-nunjucks-markdown-plugin