Update all govukHeader uses to use new crown
From the [govuk-frontend
changelog](https://github.com/alphagov/govuk-frontend/blob/main/CHANGELOG.md#update-the-logo-in-the-header-of-your-page)

Future versions of the design system will have the new crown in by
default. When that happens we can revert this commit and remove
`useTudorCrown: true` from the macro call.

This could have been done by refactoring the views to concentrate the
header calls into one place. Instead we do it literally as it's easier
and more flexible in the prototype.